Look-up mated connectors before mate processing
The code processing mates used a mix of repeated connector look-ups
and local connector variables, and one variable was used before it
was assigned the correct value.

This error was detected when investigating the #355 bug report.

The local connector variables are now both assigned initially when
processing each mate, and used instead of repeated connector look-ups.